sudoku: Fixed mark timing
The problem is, when the main window is not active(timer paused) and
the auto-save kicks in - the timer always gets restarted in mark_timing().

This patch fixes it by storing the current value of self.is_timing at the start
of mark_timing() and then checking it before calling resume_timing().

Fixes bug #562782