fix type check so gnome-sudoku works with pygobject >= 3.0.3
gnome-sudoku was using if type(grid) == str to check if it needed
to convert the game board to a list.  Unicode fixes in the latest
pygobject returns unicode strings for any string stored in a
TreeStore. The fix was to correctly check for any string using
isinstance(grid, basestring)

Note this will not work in python3 so needs to be looked at when
porting