Fixed the Sudoku crash on starting Hard/Very Hard games
According to the docs, when the MapIterator is out of track,
neither get_key, get_value, set_value nor unset are defined
and all will fail.
After the next call to next(), they will be defined again.

When the MapIterator is initialized, it is out of track.
Because of it, the assertion to check that failed,
and it caused the crash.
This patch advances the iter by calling next()

https://bugzilla.gnome.org/show_bug.cgi?id=633464