Fixed the GLib Critical error related to GLib.Timer
.continue() resumes a timer previously stopped with .stop()
But since .stop() is not called anywhere in the code,
the assertion to check that failed and raised the error.

Since we haven't implemented Pause and Resume functions,
and start_game() is meant only for starting a game,
I have removed the timer.continue() call.
When we implement those functions,
timer.stop() and timer.continue() will be used in them

https://bugzilla.gnome.org/show_bug.cgi?id=633464