Do not guess user's desired difficulty level
If the user has never played GNOME Sudoku before, or just has no games
in progress, show the new game screen when starting Sudoku rather than
badly guessing which difficulty level he might want to play.

Usually a game will be saved and launching Sudoku will take you straight
to a puzzle, but if not, one click is not a big deal.

Lastly, it'd be better to present the new game screen than an empty
puzzle, so stop saving those.

https://bugzilla.gnome.org/show_bug.cgi?id=734699