view: Increase line width
Previously the thickness of all the lines of both the number boxes and
grids were same , the only thing to differentiate between the two was
the colour difference between the lines.This was unsatisfactory hence
this commit fixes the problem by increasing the line width of the lines
drawn to specify the 9 grids.
The line width has been increased  from 1 px to 2 px using the
function set_line_width.

https://bugzilla.gnome.org/show_bug.cgi?id=749757