Require newer vala
This ensures that distros do not attempt to build without the fix for
bug #779955, as otherwise GNOME Sudoku is totally broken due to GTK+
drawing errors caused by an uninitialized PangoLayout.

With apologies to Jeremy from Ubuntu, who will not like this. ;) Note
that this bug was introduced after vala 0.34, so distros that are
sticking with vala 0.34 can safely revert this.