Handle earmarks properly in undo/redo
Before this commit, undo/redo didn't know anything about earmarks.  This
can lead to annoyances in real-world usage of the application:  Imagine
you set some earmarks on a cell and then think some of those numbers
should be the real value.  So you set the value to be what you want, but
shortly afterwards notice that it was nonsense actually.  So you undo,
but now the earmarks you set before are lost!  This is not what undo
should be like.

In addition to the value a cell has, we now simply also save the
earmarks.  Whenever you change something, the value and earmarks the
cell had before the change are saved onto the undo stack.  Whenever you
undo/redo, the earmarks are set appropriately.