Make undo and redo hide the value/earmark popovers
The bug fixed by this commit got a little bit more visible due to
"89e7bc1 Handle earmarks properly in undo/redo (shak-mar, 2019-07-09)",
because it is more apparent with the earmark popover;  however, the bug
was also present before:

1. Enter a value into a cell that was previously empty
2. Open the value popover of the cell again (Now it shows a "clear"
   button)
3. Press the undo button.  Now the value disappears from the cell, but
   the "clear" button remains, which is an inconsistent state.

One way to fix this would be to update the popovers whenever things
change underneath.  The other way, implemented by this commit, is to
simply hide the popover whenever undo/redo is called.