Display numbers at top-left in the earmark picker
Before this commit, the value and earmark picker looked completely the
same.  However, depending on which mouse button you used, you are in a
different mode, because clicking the numbers has a different effect.
It's always a good thing to make state like that visible to the user, so
now we give a slight hint by displaying the number off-center in earmark
mode.