Mobile-friendly layout - better implementation
Previous attempt (repositioning the control buttons):
https://gitlab.gnome.org/GNOME/gnome-sudoku/-/merge_requests/32

The appearance of the resulting layout was not satisfactory
and the difficulty display did not fit the headerbar.

To make the window look better buttons were moved from controls_box
(placed next to the board) to the headerbar and the primary menu.
The difficulty display was removed.
To further improve the usability on small screens clock is now being
moved to below the board when the window is too narrow.

As a consequence of the changes following entities were removed:
- main_squeezer (and libhandy dependency),
- start_box_s (smaller version of the main menu),
- aspect frames (previously used for controls placement).

https://gitlab.gnome.org/GNOME/gnome-sudoku/-/issues/38