Fix Ctrl+num not setting earmarks
When using the keyboard to set earmarks, we need to call
game.cell_changed() to actually trigger the cell to draw. Currently the
earmark picker manages this for us if it exists (by emitting the
NumberPicker::earmark_state_changed signal, which SudokuCell listens
for) but if there is no earmark picker, this step is forgotten.

Turns out there is also no need to call game.enable/disable_earmark if
the earmark picker exists, for the same reason.

Fixes #74