Fix warning about failure to solve puzzle when starting game
Since !43 we solve the puzzle when starting the game in order to show
warnings when the solution is violated. Problem is, we don't
differentiate between fixed cells and user-filled cells. We should only
consider the fixed cells when solving the puzzle.

This fixes the following warning when loading a game with any cell
entered incorrectly:

"Failed to solve puzzle: the puzzle does not have a unique solution."

This should also probably fix the puzzle consistency warnings in this
state.