number-picker: remove assumption that game is 3x3
This sort of broke in 3db33eb2d81dd674fe2fdb8205f0ce48e79ae9f4. Although
Sudoku does not support anything besides 3x3 games, the code is designed
to ensure it should work in theory. Let's avoid a 3x3 assumption that
snuck in when inverting the rows in the number picker.