Fix misuse of non-earmark number picker as an earmark picker
We cannot call NumberPicker.set_earmark() on a NumberPicker that is not
an earmark picker. This bug was introduced in
f8b33033f6bafaaa3926dc8c1c4c3472e1c66330. It results in a critical
warning when trying to add earmarks with Ctrl+Num after left-clicking on
a cell to open the number picker.

Also, add assertions to crash earlier if these methods are misused.