Manually open and close popovers
This allows us to open the new popover with the same click that closes
the previous popover, so the user doesn't have to click twice. This
mostly restores our original behavior from before the GTK 4 port. One
difference is that now clicking outside the sudoku view will cause the
popovers to be dismissed.

Fixes #79