Add Devel profile to allow use of special devel icon
!25 added a new org.gnome.Sudoku.Devel icon, but it was unused until
now.

This also fixes the name of the install desktop file and appdata to
org.gnome.Sudoku.Devel. The gsettings schema is left untouched because I
think that's not supposed to use the suffix, and the first app I checked
(Epiphany) is also not using the Devel suffix for that.

Fixes #95